concluir a tela do mapa
