+++ supportclone/New Group/Mapaa.swift
@@ -10,13 +10,12 @@ import MapKit
 import SwiftUI
 
 struct MapaView: View {
+    @State var mapRegion = MKCoordinateRegion(center: CLLocationCoordinate2D(latitude: -23.530428407997228, longitude: -46.57457693299173), span: MKCoordinateSpan(latitudeDelta: 0.2, longitudeDelta: 0.2))
     var body: some View {
         
         NavigationStack {
             ZStack {
-                @State var mapRegion = MKCoordinateRegion(center: CLLocationCoordinate2D(latitude: -23.530428407997228, longitude: -46.57457693299173), span: MKCoordinateSpan(latitudeDelta: 0.2, longitudeDelta: 0.2))
-                
-                
+                            
                 Map(coordinateRegion: $mapRegion)
                     .navigationTitle("Locais de Reparo")
                     .navigationBarTitleDisplayMode(.inline)
